// mini-app/src/types.d.ts
export interface UserInfo {
  id: string;
  name: string;
  avatar: string;
  phone: string;
  email: string;
  address: string;
}

export interface Product {
  id: number;
  name: string;
  price: number;
  originalPrice?: number;
  image: string;
  category: Category;
  detail?: string;
  sizes?: Size[];
  colors?: Color[];
  userId?: string;
}

export interface Category {
  id: number;
  name: string;
  detail: string;
  image: string;
  
  imgList?: string[]; 
  icon: string;
 
}

export interface CartItem {
  product: Product;
  quantity: number;
}

export type Cart = CartItem[];

export interface Location {
  lat: number;
  lng: number;
}

export interface ShippingAddress {
  alias: string;
  address: string;
  name: string;
  phone: string;
}

export interface Station {
  id: number;
  name: string;
  image: string;
  address: string;
  location: Location;
}

export type Delivery =
  | ({
      type: 'shipping';
    } & ShippingAddress)
  | {
      type: 'pickup';
      stationId: number;
    };

export type OrderStatus = 'pending' | 'shipping' | 'completed';
export type PaymentStatus = 'pending' | 'success' | 'failed';

export interface Order {
  id: number;
  status: OrderStatus;
  paymentStatus: PaymentStatus;
  createdAt: Date;
  receivedAt: Date;
  items: CartItem[];
  delivery: Delivery;
  total: number;
  note: string;
}